requirements.txt
pytchat
openai
python-dotenv
pydub
pyttsx3
langdetect
websocket-client
